pyproject.toml
[project]
name = "Trade_Strategy"
version = "0.1.0"
description = "Breakout strategy analysis and observation system"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "pandas-ta>=0.3.14b0",
    "ta-lib>=0.6.8",
    "pandas-stubs~=2.3.3",
    "akshare>=1.18.7",
    "scipy>=1.17.0",
    "scikit-learn>=1.8.0",
    "optuna>=4.7.0",
    "cmaes>=0.12.0",
    "yfinance>=1.2.0",
    "dash>=4.0.0",
    "zhipuai>=2.1.5.20250825",
    "finnhub-python>=2.4.27",
    "requests>=2.32.5",
    "sniffio>=1.3.1",
    "fastembed>=0.7.3",
    "openai>=2.29.0",
    "transformers>=4.57.6",
    "torch>=2.10.0",
    "tiktoken>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["BreakoutStrategy"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]
